Separated general type declaration functionality into a separate module
